Endre type og bump versjon

+++ pensjon-brevbaker-api-model/src/main/kotlin/no/nav/pensjon/brev/api/model/maler/legacy/vedtaksbrev/vedtaksdata/vilkarsvedtaklist/vilkarsvedtak/beregningsvilkar/BeregningsVilkar.kt
@@ -14,5 +14,5 @@ data class BeregningsVilkar(
     val virkningstidpunkt: LocalDate?,
     val ieuinntekt: Kroner?,
     val skadetidspunkt: LocalDate?,
-    val yrkesskadegrad: Double?
+    val yrkesskadegrad: Int?
 )
